Comments and Fixes
All different files were commented and some minor changes were carried out in the substation class. Furthermore, the net class was renamed to grid class. Lastly, the scenario parameters of the GUI was updated within the GUI code and the concurrent scenario simulations were saved within a separete script in the file filder.